# openapi_veritas.yaml
openapi: 3.1.0
info:
  title: Veritas Academic API
  version: 1.3.0
  description: API para analizar similitud textual, riesgo compatible con uso de IA, analisis combinado, almacenamiento e historial.

servers:
  - url: https://veritas-api-ujcl.onrender.com

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

  schemas:
    TextRequest:
      type: object
      required:
        - text
      properties:
        text:
          type: string
          description: Texto a analizar
        language:
          type: string
          default: es
          description: Idioma principal del texto
        title:
          type: string
          nullable: true
          description: Titulo opcional del documento

    SimilarityMatch:
      type: object
      properties:
        text_span:
          type: string
        source_title:
          type: string
        source_url:
          type: string
          nullable: true
        match_percent:
          type: number
        source_type:
          type: string
          enum:
            - web
            - paper
            - student_paper
            - internal
            - unknown
        note:
          type: string
          nullable: true

    SimilarityResponse:
      type: object
      properties:
        overall_similarity:
          type: number
        risk_level:
          type: string
          enum:
            - Bajo
            - Medio
            - Alto
            - Indeterminado
        likely_issue:
          type: string
          enum:
            - sin_hallazgos_relevantes
            - coincidencias_convencionales
            - mala_citacion_potencial
            - parafraseo_cercano_potencial
            - dependencia_fuerte_de_fuentes
            - revision_manual_urgente
        matches:
          type: array
          items:
            $ref: '#/components/schemas/SimilarityMatch'
        notes:
          type: string
        disclaimer:
          type: string

    SuspiciousSegment:
      type: object
      properties:
        text_span:
          type: string
        reason:
          type: string
        signal_strength:
          type: string

    AIRiskResponse:
      type: object
      properties:
        ai_risk_score:
          type: number
        risk_level:
          type: string
          enum:
            - Bajo
            - Medio
            - Alto
            - Indeterminado
        segments:
          type: array
          items:
            $ref: '#/components/schemas/SuspiciousSegment'
        notes:
          type: string
        disclaimer:
          type: string

    DocumentAnalysisResponse:
      type: object
      properties:
        title:
          type: string
          nullable: true
        similarity:
          $ref: '#/components/schemas/SimilarityResponse'
        ai_risk:
          $ref: '#/components/schemas/AIRiskResponse'
        overall_risk:
          type: string
          enum:
            - Bajo
            - Medio
            - Alto
            - Indeterminado
        summary:
          type: string
        conclusion:
          type: string
        recommendations:
          type: array
          items:
            type: string
        disclaimer:
          type: string

    StoredDocumentAnalysisResponse:
      allOf:
        - $ref: '#/components/schemas/DocumentAnalysisResponse'
        - type: object
          properties:
            analysis_id:
              type: integer
              nullable: true

    AnalysisListItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        overall_risk:
          type: string
          enum:
            - Bajo
            - Medio
            - Alto
            - Indeterminado
          nullable: true
        similarity_score:
          type: number
          nullable: true
        ai_risk_score:
          type: number
          nullable: true
        created_at:
          type: string
          format: date-time

    AnalysisDetailResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
          nullable: true
        language:
          type: string
          nullable: true
        text:
          type: string
        similarity_score:
          type: number
          nullable: true
        similarity_risk:
          type: string
          nullable: true
        ai_risk_score:
          type: number
          nullable: true
        ai_risk_level:
          type: string
          nullable: true
        overall_risk:
          type: string
          enum:
            - Bajo
            - Medio
            - Alto
            - Indeterminado
          nullable: true
        summary:
          type: string
          nullable: true
        conclusion:
          type: string
          nullable: true
        recommendations:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time

paths:
  /health:
    get:
      operationId: healthCheck
      summary: Verifica estado de la API
      security: []
      responses:
        '200':
          description: API activa
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  environment:
                    type: string

  /analyze_document:
    post:
      operationId: analyzeDocument
      summary: Ejecuta un analisis combinado de similitud, riesgo IA y conclusion general
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextRequest'
      responses:
        '200':
          description: Resultado combinado del analisis del documento
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StoredDocumentAnalysisResponse'
        '401':
          description: No autorizado

  /check_similarity:
    post:
      operationId: checkSimilarity
      summary: Analiza similitud textual y coincidencias relevantes
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextRequest'
      responses:
        '200':
          description: Resultado de similitud
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimilarityResponse'
        '401':
          description: No autorizado

  /check_ai_risk:
    post:
      operationId: checkAiRisk
      summary: Estima senales compatibles con uso de IA
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TextRequest'
      responses:
        '200':
          description: Resultado de riesgo IA
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIRiskResponse'
        '401':
          description: No autorizado

  /analyses:
    get:
      operationId: listAnalyses
      summary: Lista analisis guardados
      security:
        - BearerAuth: []
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Lista de analisis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnalysisListItem'
        '401':
          description: No autorizado
        '503':
          description: Base de datos no disponible

  /analyses/{analysis_id}:
    get:
      operationId: getAnalysisById
      summary: Obtiene un analisis guardado por ID
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: analysis_id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Analisis encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisDetailResponse'
        '401':
          description: No autorizado
        '404':
          description: Analisis no encontrado
        '503':
          description: Base de datos no disponible